+++ client/src/api/User/index.js
@@ -3,3 +3,4 @@ export * from './signUp';
 export * from './modify';
 export * from './logout';
 export * from './refresh';
+export * from './activity';
