+++ client/src/pages/PostPage/Answer.js
@@ -7,7 +7,7 @@ import AvatarWrapper from '../../components/AvatarWrapper';
 import MessageIcon from '@mui/icons-material/Message';
 import { ISOHandler } from '../../utils/timeHandler';
 import { getUserId } from '../../utils/cookies';
-import { deleteAnswer, updateAnswerVotes } from '../../utils/axiosRequest';
+import { deleteAnswer, updateAnswerVotes } from '../../api/Answer';
 import { answerStore } from '../../store/store';
 import AlertDialog from '../../components/AlertDialog';
 import { useParams } from 'react-router-dom';
