+++ client/src/pages/UserPage/UserPage.js
@@ -1,15 +1,11 @@
 import { Box, Button, Tabs, Tab } from '@mui/material';
-
 import UserPageAvatarWrapper from './UserPageAvatarWrapper';
 import styled from 'styled-components';
 import BasicTabs from '../../components/Tab/BasicTabs';
 import { useState, useEffect } from 'react';
 import AnswerComment from './AnswerComment';
 import ModifyUserPage from './ModifyUserPage';
-import {
-  getAccessWithRefresh,
-  getUsersActivity,
-} from '../../utils/axiosRequest';
+import { getAccessWithRefresh, getUsersActivity } from '../../api/User';
 import Pagination from '../../components/Pagination';
 import { userStore } from '../../store/store';
 import { getAccessToken, getUserId } from '../../utils/cookies';
